included reference to add user form
I created a work in progress form to let useres generate their csv entry themselves and have a basic verification process.

its in a different repository for now, to not confuse anyone forking this repo, but I intendt to include it in the long run.

Repo here, if you want to try it out on your own: https://github.com/trutzig89182/AoM-add-user