change selectAll button to select current entries
With the added filtering option the "select all" button could be a bit confusing as it also selects all accouts that are not visible when a specific keyword is active.

This commit changes it, so that only the currently visible accounts are selected, when select all is clicked.

For me this seems more intuitive, but is it?

relates to #15