Add basic support to allow filtering by language
relates to #15 #7

Adds languages to webpage and allows to filter list by clicking on them.
However there are still some problems: filtering by keywords and by languages should work in a complementary manner, so that while a language is selected only entries for this language are shown if a keyword is selected in addition. right now, selecting a keyword will just overwrite the selection based on langue.