Change "select all" button to only select all visible elements (#67)
* Add handling of empty profile url

If "link" in csv is empty, name will be redered as plain text. An error message is written to the log.

* Update .gitignore

* change selectAll button to select current entries

With the added filtering option the "select all" button could be a bit confusing as it also selects all accouts that are not visible when a specific keyword is active.

This commit changes it, so that only the currently visible accounts are selected, when select all is clicked.

For me this seems more intuitive, but is it?

relates to #15

* avoid build list to break when no "keywords" in CSV file

The building of the list could not be completed without a column "keywords" being in the CSV list. This is fixed here, by not adding keywords when users.keywords is undefined. resolves #64